rsn
Function to_writer

pub fn to_writer<S: Serialize, W: Write>(value: &S, writer: W) -> Result<usize>

Serializes value into a writer using the default Rsn Config.

let mut serialized = Vec::new();
rsn::to_writer(&vec![1, 2, 3], &mut serialized).unwrap();
assert_eq!(serialized, b"[1,2,3]");

Errors

Returns any errors occurring while serializing value or while writing to writer.