rsn::writer
Struct Writer

pub struct Writer<'config, Output> { /* private fields */ }

A low-level writer for the Rsn format.

Implementations

impl<'config, Output> Writer<'config, Output>
where Output: Write,

pub fn new(output: Output, config: &'config Config) -> Self

Returns a writer that outputs to output using config.

pub fn finish(self) -> Output

Finishes writing and returns the output.

Panics

This function panics if a nested type was begun and a corresponding finish_nested was not called.

pub fn begin_named_map(&mut self, name: &str) -> Result

Begins a named map. A corresponding call to finish_nested must be made when the map contents are completed.

Errors

Returns any errors that arise while writing to Output.

pub fn begin_named_tuple(&mut self, name: &str) -> Result

Begins a named tuple. A corresponding call to finish_nested must be made when the tuple contents are completed.

Errors

Returns any errors that arise while writing to Output.

pub fn begin_map(&mut self) -> Result

Begins a map. A corresponding call to finish_nested must be made when the map contents are completed.

Errors

Returns any errors that arise while writing to Output.

pub fn begin_tuple(&mut self) -> Result

Begins a tuple. A corresponding call to finish_nested must be made when the tuple contents are completed.

Errors

Returns any errors that arise while writing to Output.

pub fn begin_list(&mut self) -> Result

Begins a list/array. A corresponding call to finish_nested must be made when the list contents are completed.

Errors

Returns any errors that arise while writing to Output.

pub fn write_primitive<P>(&mut self, p: &P) -> Result
where P: Primitive + ?Sized,

Writes a primitive value, formatting it as valid Rsn.

Errors

Returns any errors that arise while writing to Output.

pub fn write_raw_value(&mut self, ident: &str) -> Result

Writes ident without any extra formatting.

Errors

Returns any errors that arise while writing to Output.

pub fn insert_newline(&mut self) -> Result

Inserts the configured newline character, if needed.

Errors

Returns any errors that arise while writing to Output.

pub fn finish_nested(&mut self) -> Result

Finishes the current nested value.

Errors

Returns any errors that arise while writing to Output.

Panics

This function panics if are no open nested types.

pub fn write_value(&mut self, value: &Value<'_>) -> Result

Writes a value.

Errors

Returns any errors that arise while writing to Output.