rsn::writer
Trait Primitive

pub trait Primitive {
    // Required method
    fn render_to<W: Write>(&self, buffer: &mut W) -> Result;
}

A type that can be written as a primitive.q

Required Methods

fn render_to<W: Write>(&self, buffer: &mut W) -> Result

Renders this type to buffer.

Errors

Returns any errors that arise while writing to buffer.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.